added dummy data objects

+++ src/main/scala/models/Product.scala
@@ -10,9 +10,17 @@ import schema.SchemaUtils.productSchema
   * @param name - product name
   * @param price - product price
   */
-case class Product(id: String, name: String, price: Float)
+case class  Product(id: String, name: String, price: Double)
 
 object Product {
+
+  /**
+    * Creates Dataset of type Product
+    * @param sparkSession - Spark Session
+    * @param topic - Kafka topic
+    * @param productEncoderEncoder - Encoder of type Product
+    * @return - Dataset of type Product
+    */
   def getProductDS(sparkSession: SparkSession, topic: String)(implicit productEncoderEncoder: Encoder[Product]): Dataset[Product] = {
     createSource[Product](sparkSession, topic, (df: DataFrame) => df.as[Product])
   }
